refactor(app): move dynamic to add component in separate file
